refactor: split mcp methods.ts into a methods/ folder by feature
methods.ts had grown to ~1000 lines. Split it into src/editor/mcp/methods/,
one file per feature domain, mirroring the editor-mcp-server tool-file layout
(entity/asset/scene/project/viewport/runtime/store/editor/vcs). Shared helpers
(api, log, rest, iterateObject, entitySummary, paginate) move to methods/shared.ts;
domain-only helpers stay with their domain. methods-vcs.ts moves in as methods/vcs.ts.

Pure move, no behavior change: all 69 mcp.method registrations preserved, tsc and
eslint clean. index.ts imports the folder via its barrel (import './mcp/methods'
resolves to methods/index.ts), so the editor entry only drops the separate
methods-vcs import.
