+++ src/editor/mcp/methods/index.ts
@@ -1,0 +1,9 @@
+import './entity';
+import './asset';
+import './scene';
+import './project';
+import './viewport';
+import './runtime';
+import './store';
+import './editor';
+import './vcs';
